fix(solve): continue fork PRs instead of halting on permission-denied push (#1893)
When continuing a cross-repository PR from another contributor's fork with
'Allow edits by maintainers' enabled, the solver synced the upstream default
branch and pushed it to origin (the contributor's fork). The maintainer does
not own that fork, so GitHub rejected the push with 'permission denied', which
was misclassified as fork divergence and halted the run.

- Add isPermissionDeniedPushError() and shouldPushDefaultBranchToFork() pure
  helpers in solve.branch-divergence.lib.mjs.
- Extract setupUpstreamAndSync into solve.fork-sync.lib.mjs (re-exported) to
  stay under the 1500-line limit; skip the fork default-branch push when the
  user does not own the fork, and treat permission-denied as non-fatal.
- Add regression tests and a deep case study under docs/case-studies/issue-1893.
